backend/rss-service/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.4
feedparser==6.0.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
playwright==1.48.0
psycopg-binary==3.2.3
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
python-dotenv==1.0.1
requests==2.32.3
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
confluent-kafka==2.6.0